Refactor grafico function
